centralize the common types
